added utils.js and integrated with a couple hacks
not a comprehensive integration; this is meant as a quick demo of how
import works
