+++ src/custom text effect.js
@@ -40,6 +40,8 @@ The second argument is `time`, which is the time in milliseconds
 
 A number of example effects are included
 */
+import {inject} from "./utils.js";
+
 var customTextEffects = {
 	"my-effect": function () {
 		// a horizontal wavy effect with a blue tint 
@@ -112,39 +114,6 @@ var customTextEffects = {
 			char.offset.y -= Math.pow(char.col - lastSpace, 1.5) * (Math.sin(time / 120 + char.col / 2));
 		}
 	},
-};
-
-// helper used to inject code into script tags based on a search string
-var inject = function (searchString, codeToInject) {
-	// find the relevant script tag
-	var scriptTags = document.getElementsByTagName('script');
-	var scriptTag;
-	var code;
-	for (var i = 0; i < scriptTags.length; ++i) {
-		scriptTag = scriptTags[i];
-		if (
-			scriptTag.textContent.indexOf(searchString) >= 0 // script contains the search string
-			&&
-			scriptTag != document.currentScript // script isn't the one doing the injecting (which also contains the search string)
-		) {
-			code = scriptTag.textContent;
-			break;
-		}
-	}
-
-	// error-handling
-	if (!code) {
-		throw 'Couldn\'t find "' + searchString + '" in script tags';
-	}
-
-	// modify the content
-	code = code.replace(searchString, searchString + codeToInject);
-
-	// replace the old script tag with a new one using our modified code
-	scriptTag.remove();
-	scriptTag = document.createElement('script');
-	scriptTag.textContent = code;
-	document.head.appendChild(scriptTag);
 };
 
 // generate code for each text effect
